+++ O3/dappbrowserv2/dAppBrowserV2ViewController.swift
@@ -26,163 +26,8 @@ protocol dAppBrowserDelegate {
     func beginLoading()
 }
 
-class dAppBrowserViewModel: NSObject {
-    
-    var isConnected: Bool = false
-    var connectedTime: Date?
-    var url: URL!
-    var delegate: dAppBrowserDelegate?
-    var dappMetadata: dAppMetadata? = dAppMetadata()
-    var selectedAccount: NEP6.Account?
-    var unlockedWallet: Wallet?
-    var assetSymbol: String? = nil
-    var tradingAccount: TradingAccount? = nil
-    var tradableAsset: TradableAsset? = nil
-    
-    func loadMetadata(){
-        OpenGraph.fetch(url: url!) { og, error in
-            self.dappMetadata?.url = self.url!
-            self.dappMetadata?.title = og?[.title]
-            self.dappMetadata?.iconURL = og?[.image]
-            self.dappMetadata?.description = og?[.description]
-        }
-    }
-    
-    func requestToConnect(message: dAppMessage, didCancel: @escaping (_ message: dAppMessage) -> Void, didConfirm: @escaping (_ message: dAppMessage, _ wallet: Wallet, _ acount: NEP6.Account?) -> Void) {
-        
-        self.delegate?.onConnectRequest(url: self.url, message: message, didCancel: { m in
-            didCancel(m)
-        }, didConfirm: { m, wallet, account in
-            self.isConnected = true
-            self.connectedTime = Date()
-            self.unlockedWallet = wallet
-            self.selectedAccount = account
-            didConfirm(m, wallet, account)
-        })
-    }
-    
-    func requestToSend(message: dAppMessage, request: dAppProtocol.SendRequest, didCancel: @escaping (_ message: dAppMessage,_ request: dAppProtocol.SendRequest) -> Void, onCompleted:@escaping (_ response: dAppProtocol.SendResponse?, _ error: dAppProtocol.errorResponse?) -> Void) {
-        self.delegate?.onSendRequest(message: message, request: request, didCancel: didCancel, onCompleted: onCompleted)
-    }
-    
-    func responseWithError(message: dAppMessage, error: String) {
-        self.delegate?.error(message: message, error: error)
-    }
-    
-    func proceedMessage(message: dAppMessage) {
-        
-        if message.command.lowercased() == "getAccount".lowercased() {
-            if unlockedWallet == nil {
-                
-                return
-            }
-            let response = dAppProtocol.GetAccountResponse(address: unlockedWallet!.address, publicKey: unlockedWallet!.publicKeyString)
-            self.delegate?.didFinishMessage(message: message, response: response.dictionary)
-            return
-        }
-        
-        if message.command.lowercased() == "getProvider".lowercased() {
-            let response = dAppProtocol.GetProviderResponse(name: "o3", version: "v1", website: "https://o3.network", compatibility: ["NEP-dapi"])
-            self.delegate?.didFinishMessage(message: message, response: response.dictionary)
-            return
-        }
-        
-        if message.command.lowercased() == "getNetworks".lowercased() {
-            let response = dAppProtocol.GetNetworksResponse(networks: ["MainNet", "TestNet", "PrivateNet"])
-            self.delegate?.didFinishMessage(message: message, response: response.dictionary)
-            return
-        }
-        
-        if message.command.lowercased() == "getBalance".lowercased() {
-            //parse input
-            let decoder = JSONDecoder()
-            guard let dictionary =  message.data?.value as? JSONDictionary,
-                let data = try? JSONSerialization.data(withJSONObject: dictionary, options: .prettyPrinted),
-                let request = try? decoder.decode(dAppProtocol.RequestData<[dAppProtocol.GetBalanceRequest]>.self, from: data) else {
-                    self.delegate?.error(message: message, error: "Unable to parse the request")
-                    return
-            }
-             self.delegate?.beginLoading()
-            DispatchQueue.global().async {
-            let response = O3DappAPI().getBalance(request: request)
-            //it is very important to make the struct to dictionary otherwise JSONDecoder will throw and error invalid SwiftValue when trying to decode it
-            self.delegate?.didFinishMessage(message: message, response: response.dictionary)
-            
-            }
-            return
-        }
-        
-        if message.command.lowercased() == "getStorage".lowercased() {
-            let decoder = JSONDecoder()
-            guard let dictionary =  message.data?.value as? JSONDictionary,
-                let data = try? JSONSerialization.data(withJSONObject: dictionary, options: .prettyPrinted),
-                let request = try? decoder.decode(dAppProtocol.GetStorageRequest.self, from: data) else {
-                    self.delegate?.error(message: message, error: "Unable to parse the request")
-                    return
-            }
-            let response = O3DappAPI().getStorage(request: request)
-            self.delegate?.didFinishMessage(message: message, response: response.dictionary)
-            return
-        }
-        
-        if message.command.lowercased() == "invokeRead".lowercased() {
-            let decoder = JSONDecoder()
-            guard let dictionary =  message.data?.value as? JSONDictionary,
-                let data = try? JSONSerialization.data(withJSONObject: dictionary, options: .prettyPrinted),
-                let request = try? decoder.decode(dAppProtocol.InvokeReadRequest.self, from: data) else {
-                    self.delegate?.error(message: message, error: "Unable to parse the request")
-                    return
-            }
-            let response = O3DappAPI().invokeRead(request: request)
-            self.delegate?.didFinishMessage(message: message, response: response)
-            return
-        }
-        
-        if message.command.lowercased() == "invoke".lowercased() {
-            let decoder = JSONDecoder()
-            guard let dictionary =  message.data?.value as? JSONDictionary,
-                let data = try? JSONSerialization.data(withJSONObject: dictionary, options: .prettyPrinted),
-                let request = try? decoder.decode(dAppProtocol.InvokeRequest.self, from: data) else {
-                    self.delegate?.error(message: message, error: "Unable to parse the request")
-                    return
-            }
-            let response = O3DappAPI().invoke(request: request)
-            self.delegate?.didFinishMessage(message: message, response: response.dictionary)
-            return
-        }
-        
-        if message.command.lowercased() == "send".lowercased() {
-            let decoder = JSONDecoder()
-            guard let dictionary =  message.data?.value as? JSONDictionary,
-                let data = try? JSONSerialization.data(withJSONObject: dictionary, options: .prettyPrinted),
-                var request = try? decoder.decode(dAppProtocol.SendRequest.self, from: data) else {
-                    self.delegate?.error(message: message, error: "Unable to parse the request")
-                    return
-            }
-            request.fromAddress = unlockedWallet!.address
-            self.requestToSend(message: message, request: request, didCancel: { m,r in
-                self.delegate?.error(message: message, error: "USER_CANCELLED_SEND")
-            }, onCompleted: { response, err in
-                self.delegate?.beginLoading()
-                DispatchQueue.global().async {
-                    self.delegate?.didFinishMessage(message: message, response: response!.dictionary)
-                }
-            })
-            
-            return
-        }
-    }
-    
-    func changeActiveAccount(account: NEP6.Account? ,wallet: Wallet) {
-        self.unlockedWallet = wallet
-        self.selectedAccount = account
-        self.delegate?.onWalletChanged(newWallet: wallet)
-    }
-}
 
 class dAppBrowserV2ViewController: UIViewController {
-    
-    
     @IBOutlet weak var containerView: UIView!
     @IBOutlet weak var backButton: UIBarButtonItem!
     @IBOutlet weak var walletSwitcherButton: UIBarButtonItem!
@@ -198,34 +43,13 @@ class dAppBrowserV2ViewController: UIViewController {
     func addObservers() {
         NotificationCenter.default.addObserver(self, selector: #selector(loadOpenOrders), name: NSNotification.Name(rawValue: "needsReloadOpenOrders"), object: nil)
         NotificationCenter.default.addObserver(self, selector: #selector(viewOpenOrders(_:)), name: NSNotification.Name(rawValue: "viewTradingOrders"), object: nil)
-    
     }
     
-    
-    @objc private func loadOpenOrders() {
-        O3APIClient(network: AppState.network).loadSwitcheoOrders(address: Authenticated.wallet!.address, status: SwitcheoOrderStatus.open) { result in
-            switch result{
-            case .failure(let error):
-                #if DEBUG
-                print(error)
-                #endif
-            case .success(let response):
-                DispatchQueue.main.async {
-                    self.openOrderButton?.isHidden =  response.switcheo.count == 0
-                    self.openOrderButton?.badgeValue = String(format: "%d",response.switcheo.count)
-                }
-            }
-        }
-    }
-    
-    @objc @IBAction func viewOpenOrders(_ sender: Any) {
-        guard let nav = UIStoryboard(name: "Trading", bundle: nil).instantiateViewController(withIdentifier: "OrdersTabsViewControllerNav") as? UINavigationController else {
-            return
-        }
-        let transitionDelegate = DeckTransitioningDelegate()
-        nav.transitioningDelegate = transitionDelegate
-        nav.modalPresentationStyle = .custom
-        self.present(nav, animated: true, completion: nil)
+    deinit {
+        self.webView.removeObserver(self, forKeyPath: "estimatedProgress", context: nil)
+        self.webView.removeObserver(self, forKeyPath: "loading", context: nil)
+        NotificationCenter.default.removeObserver(self, name: Notification.Name(rawValue: "needsReloadOpenOrders"), object: nil)
+        NotificationCenter.default.removeObserver(self, name: Notification.Name(rawValue: "viewTradingOrders"), object: nil)
     }
     
     
@@ -269,31 +93,6 @@ class dAppBrowserV2ViewController: UIViewController {
         }
     }
     
-    @objc private func loadTradingAccountBalances() {
-        O3APIClient(network: AppState.network).tradingBalances(address: Authenticated.wallet!.address) { result in
-            switch result {
-            case .failure(let error):
-                print(error)
-                return
-            case .success(let tradingAccount):
-                DispatchQueue.main.async {
-                    self.viewModel.tradingAccount = tradingAccount
-                }
-            }
-        }
-    }
-    
-    private  func loadTradableAssets(completion: @escaping ([TradableAsset]) -> Void) {
-        O3APIClient.shared.loadSupportedTokenSwitcheo { result in
-            switch result {
-            case .failure(let error):
-                print(error)
-            case .success(let response):
-                completion(response)
-            }
-        }
-    }
-    
     @IBAction func tradeTapped(_ sender: Any) {
         if (viewModel.assetSymbol?.lowercased() != "neo") {
             self.showActionSheetAssetInTradingAccount(asset: self.viewModel.tradableAsset!)
@@ -301,78 +100,7 @@ class dAppBrowserV2ViewController: UIViewController {
             showBuyOptionsNEO()
         }
     }
-    
-    func openCreateOrder(action: CreateOrderAction, asset: TradableAsset) {
-        let nav = UIStoryboard(name: "Trading", bundle: nil).instantiateViewController(withIdentifier: "CreateOrderTableViewControllerNav") as! UINavigationController
-        if let vc = nav.viewControllers.first as? CreateOrderTableViewController {
-            vc.viewModel = CreateOrderViewModel()
-            vc.viewModel.selectedAction = action
-            let inTradingAccount = viewModel.tradingAccount?.switcheo.confirmed.first(where: { t -> Bool in
-                return t.symbol.uppercased() == asset.symbol.uppercased()
-            })
-            vc.viewModel.wantAsset = inTradingAccount != nil ? inTradingAccount : asset
-            vc.viewModel.offerAsset = viewModel?.tradingAccount?.switcheo.basePairs.filter({ t -> Bool in
-                return t.symbol != asset.symbol
-            }).first
-            vc.viewModel.tradingAccount = viewModel.tradingAccount
-        }
-        self.present(nav, animated: true, completion: nil)
-    }
-    
-    func showActionSheetAssetInTradingAccount(asset: TradableAsset) {
-        
-        let alert = UIAlertController(title: asset.name, message: nil, preferredStyle: .actionSheet)
-        
-        let buyButton = UIAlertAction(title: "Buy", style: .default) { _ in
-            tradingEvent.shared.startBuy(asset: asset.symbol, source: TradingActionSource.tokenDetail)
-            self.openCreateOrder(action: CreateOrderAction.Buy, asset: asset)
-        }
-        alert.addAction(buyButton)
-        
-        //we can't actually sell NEO but rather use NEO to buy other asset
-        if asset.symbol != "NEO" {
-            let sellButton = UIAlertAction(title: "Sell", style: .default) { _ in
-                tradingEvent.shared.startSell(asset: asset.symbol, source: TradingActionSource.tokenDetail)
-                self.openCreateOrder(action: CreateOrderAction.Sell, asset: asset)
-            }
-            alert.addAction(sellButton)
-        }
-        
-        
-        let cancel = UIAlertAction(title: OzoneAlert.cancelNegativeConfirmString, style: .cancel) { _ in
-            
-        }
-        alert.addAction(cancel)
-        present(alert, animated: true, completion: nil)
-    }
-    
-    func showBuyOptionsNEO() {
-        let actionSheet = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
-        let buyWithFiat = UIAlertAction(title: "With Fiat", style: .default) { _ in
-            Controller().openDappBrowserV2(url: URL(string: "https://buy.o3.network/?a=" + (Authenticated.wallet?.address)!)!)
-        }
-        actionSheet.addAction(buyWithFiat)
-        
-        let buyWithCrypto = UIAlertAction(title: "With Crypto", style: .default) { _ in
-            Controller().openDappBrowserV2(url: URL(string: "https://o3.network/swap/")!)
-        }
-        actionSheet.addAction(buyWithCrypto)
-        
-        let cancel = UIAlertAction(title: OzoneAlert.cancelNegativeConfirmString, style: .cancel) { _ in
-            
-        }
-        actionSheet.addAction(cancel)
-        present(actionSheet, animated: true, completion: nil)
-    }
-    
-    
-    deinit {
-        self.webView.removeObserver(self, forKeyPath: "estimatedProgress", context: nil)
-        self.webView.removeObserver(self, forKeyPath: "loading", context: nil)
-        NotificationCenter.default.removeObserver(self, name: Notification.Name(rawValue: "needsReloadOpenOrders"), object: nil)
-        NotificationCenter.default.removeObserver(self, name: Notification.Name(rawValue: "viewTradingOrders"), object: nil)
-    }
-    
+
     override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey : Any]?, context: UnsafeMutableRawPointer?) {
         
         if (keyPath == "estimatedProgress") {
@@ -458,6 +186,8 @@ class dAppBrowserV2ViewController: UIViewController {
             urlComps.queryItems = queryItems as [URLQueryItem]
             req = URLRequest(url: urlComps.url!)
         }
+        
+        openOrderButton.addTarget(self, action: #selector(viewOpenOrders(_:)), for: UIControl.Event.touchUpInside)
         
         self.viewModel.loadMetadata()
         self.webView.load(req)
